requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
facenet-pytorch==2.5.2
Flask==1.1.2
h5py==3.2.1
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras==2.4.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mtcnn==0.1.0
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
scipy==1.6.1
six==1.15.0
torch==1.8.0
torchvision==0.9.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
